requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.41.0
Django==5.1.14
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
drf-nested-routers==0.94.1
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
gunicorn
channels==4.1.0
daphne==4.1.2
